refactor: Utilities
Small refactor to move all utilities into a dedicated utilities directory, and ensure that anything exported from that dir is exported from the library itself as well
