+++ lib/AlternativeNamesFieldArray/AlternativeNamesField.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Field } from 'react-final-form';
 
 import { TextField } from '@folio/stripes/components';
-import { required } from '../validators';
+import { requiredValidator as required } from '../utils';
 
 export default class AlternativeNamesField extends React.Component {
   static propTypes = {
