+++ lib/DocumentFilter/DocumentFilterRule.js
@@ -12,7 +12,7 @@ import {
 } from '@folio/stripes/components';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { Field } from 'react-final-form';
-import { required as requiredValidator } from '../validators';
+import { requiredValidator } from '../utils';
 import DocumentFilterRuleConstants from './DocumentFilterRuleConstants';
 
 const DocumentFilterRule = ({
