+++ lib/DocumentsFieldArray/DocumentField.js
@@ -13,7 +13,7 @@ import {
 } from '@folio/stripes/components';
 
 import FileUploaderField from '../FileUploaderField';
-import { composeValidators } from '../validators';
+import { composeValidators } from '../utils';
 
 const MAX_URL_SIZE = 2048;
 export default class DocumentField extends React.Component {
