+++ lib/EResourceType/EResourceType.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import isPackage from '../isPackage';
+import { isPackage } from '../utils';
 
 export default class EResourceType extends React.Component {
   static propTypes = {
