+++ lib/InternalContactCard/InternalContactCard.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
+
 import { Link } from 'react-router-dom';
 import { AppIcon } from '@folio/stripes/core';
-import { FormattedMessage } from 'react-intl';
 import { Card, Col, KeyValue, NoValue, Row } from '@folio/stripes/components';
 
-import renderUserName from '../renderUserName';
+import { renderUserName } from '../utils';
 
 export default class ContactCard extends React.Component {
   static propTypes = {
