+++ lib/InternalContactSelection/InternalContactSelection.js
@@ -1,9 +1,10 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import uniqBy from 'lodash/uniqBy';
 
 import { useChunkedUsers, useParallelBatchFetch } from '../hooks';
-import renderUserName from '../renderUserName';
+import { renderUserName } from '../utils';
+
 import InternalContactSelectionDisplay from './InternalContactSelectionDisplay';
 
 const InternalContactSelection = ({
