+++ lib/InternalContactsArrayDisplay/InternalContactsArrayDisplay.js
@@ -4,7 +4,7 @@ import { useQuery } from 'react-query';
 import { useOkapiKy } from '@folio/stripes/core';
 import { Row } from '@folio/stripes/components';
 
-import renderUserName from '../renderUserName';
+import { renderUserName } from '../utils';
 
 const InternalContactsArrayDisplay = ({ contacts }) => {
   /* eslint-disable no-param-reassign */
