+++ lib/InternalContactsFieldArray/InternalContactsFieldArray.js
@@ -12,7 +12,7 @@ import {
 
 import { useKiwtFieldArray } from '@k-int/stripes-kint-components';
 import EditCard from '../EditCard';
-import { composeValidators, required } from '../validators';
+import { composeValidators, requiredValidator as required } from '../utils';
 
 import UserField from './UserField';
