+++ lib/InternalContactsFieldArray/UserField.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { Link } from 'react-router-dom';
+
 import {
   Button,
   Card,
@@ -13,9 +15,9 @@ import {
 } from '@folio/stripes/components';
 import { AppIcon, Pluggable } from '@folio/stripes/core';
 
-import { Link } from 'react-router-dom';
+import { renderUserName } from '../utils';
+
 import css from '../styles.css';
-import renderUserName from '../renderUserName';
 
 export default class UserField extends React.Component {
   static propTypes = {
