+++ lib/OrganizationsFieldArray/EditOrganizationCard.js
@@ -12,7 +12,7 @@ import {
 } from '@folio/stripes/components';
 
 import EditCard from '../EditCard';
-import { required } from '../validators';
+import { requiredValidator as required } from '../utils';
 import ViewOrganizationCard from '../ViewOrganizationCard';
 import RolesFieldArray from './RolesFieldArray';
 import FieldPrimaryOrg from './FieldPrimaryOrg';
