+++ lib/OrganizationsFieldArray/RolesFieldArray.js
@@ -14,7 +14,7 @@ import {
 } from '@folio/stripes/components';
 import { useKiwtFieldArray } from '@k-int/stripes-kint-components';
 
-import { required } from '../validators';
+import { requiredValidator as required } from '../utils';
 import { usePrevious } from '../hooks';
 
 // Utility function to check if two arrays of scalars contain the same items (Order does not count)
