+++ lib/Tags/Tags.js
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 import PropTypes from 'prop-types';
 
 import { FormattedMessage } from 'react-intl';
@@ -12,7 +12,7 @@ import { Pane } from '@folio/stripes/components';
 import { TagsForm } from '@folio/stripes/smart-components';
 
 import { useTags } from '../hooks';
-import { tagNamespaceArray } from '../tagsConfig';
+import { tagNamespaceArray } from '../utils';
 
 const Tags = ({
   invalidateLinks = [], // If there are other queries that need invalidating, pass those here
