+++ lib/hooks/useExportLogStream/useExportLogStream.js
@@ -4,7 +4,7 @@ import { useQuery } from 'react-query';
 
 import { CalloutContext, useOkapiKy } from '@folio/stripes/core';
 
-import downloadBlob from '../../downloadBlob';
+import { downloadBlob } from '../../utils';
 
 const useExportLogStream = (job, type) => {
   const callout = useContext(CalloutContext);
