+++ lib/hooks/useFileHandlers.js
@@ -1,7 +1,7 @@
 import { useMutation } from 'react-query';
 
 import { useOkapiKy } from '@folio/stripes/core';
-import downloadBlob from '../downloadBlob';
+import { downloadBlob } from '../utils';
 
 const useFileHandlers = (fileEndpoint) => {
   const ky = useOkapiKy();
