+++ lib/hooks/useTags.js
@@ -1,6 +1,6 @@
 import { useQuery } from 'react-query';
 import { useOkapiKy } from '@folio/stripes/core';
-import { defaultTagQuery, tagNamespaceArray } from '../tagsConfig';
+import { defaultTagQuery, tagNamespaceArray } from '../utils';
 
 const useTags = (namespaceArray, options) => {
   const ky = useOkapiKy();
