+++ lib/utils/getResourceIdentifier/getResourceIdentifier.js
