+++ lib/utils/getResourceIdentifier/getResourceIdentifier.test.js
