+++ lib/utils/getResourceIdentifier/index.js
