+++ lib/utils/getResourceIdentifier/testResources.js
