+++ lib/utils/getSASParams.js
