+++ lib/utils/getSiblingIdentifier/getSiblingIdentifier.js
