+++ lib/utils/getSiblingIdentifier/getSiblingIdentifier.test.js
