+++ lib/utils/getSiblingIdentifier/index.js
