+++ lib/utils/getSiblingIdentifier/testResources.js
