+++ lib/utils/isPackage.js
