+++ lib/utils/keyboardShortcutHandlers.js
