+++ lib/utils/preventResourceRefresh.js
