+++ lib/utils/renderUserName.js
