+++ lib/utils/tagsConfig.js
